feat: 사이드바 모바일 대응
